requirements.txt
pytz
python-dateutil
django
numpy
pandas
scipy
websocket-client
requests
channels
whitenoise
psycopg2.binary
